api: test: guarantee a bad signature when checking for it
NIBPutTest.TestPutMalformedSignature attempted to check
whether a corrupted signature would lead to an error
as expected.
It did this by overwriting one byte of the signature
with 0.
As 0 can be a legitimate of a signature, 1/256 of all
test runs this test did not yield the expected failure,
because the random signature really contained a 0 at
the position in question.

We overwrite the full signature with 0s now, effectively
eliminating the chances for this to happen.

See #83.